lib/rfsv.h: make the const variable RFSV_SENDLEN inline
This makes it extern, so we don’t get a warning about unused variables
in translation units that don’t use it, doesn’t force us to define it in a
separate file (as making it just extern would), and retains the typechecking
of a proper variable (not a #define).

+++ sisinstall/sisinstaller.cpp
@@ -256,12 +256,12 @@ SISInstaller::installFile(SISFileRecord* fileRecord)
                 case 3:
                         if (logLevel >= 1)
                                 fprintf(stderr, "Run %.*s during installation/remove\n",
-                                                fileRecord->m_destLength, fileRecord->getDestPtr());
+                                        (int) fileRecord->m_destLength, fileRecord->getDestPtr());
                         break;
                 case 4:
                         if (logLevel >= 2)
                                 fprintf(stderr, "Running the app will create %.*s\n",
-                                                fileRecord->m_destLength, fileRecord->getDestPtr());
+                                        (int) fileRecord->m_destLength, fileRecord->getDestPtr());
                         break;
                 }
         return FILE_OK;
